\name{aboutKTS}
\alias{aboutKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
aboutKTS: opens information file about KarsTS version
}

\description{
This function opens the information file about KarsTS version
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
