\name{applyGap2TSer}
\alias{applyGap2TSer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
applyGap2TSer: apply a gap set to a time series
}
\description{
This function applies a set of gaps to a time series.
}
%- maybe also 'usage' for other objects documented here.
\details{
The new time series name is a combination of the original time series and the gap set names.
}
\value{
A new time series with NAs at the locations indicated by the gap set
}
\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
