\name{buttons1}
\alias{buttons1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
buttons1: create the buttons corresponding to the Time Series Menu
}

\description{
This function creates the buttons corresponding to the Time Series Menu
}

\author{
Marina Saez Andreu
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
