\name{buttons3}
\alias{buttons3}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
buttons3: create the buttons corresponding to the Analysis Menu
}

\description{
This function creates the buttons corresponding to the Analysis Menu
}

\author{
Marina Saez Andreu
}
