\name{checkIfAny}
\alias{checkIfAny}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
checkIfAny: check if there are any data sets in the environment
}
\description{
This functions checks whether there are any data sets (time series, gap sets or recurrence matrices) in the environment (KTSEnv). If it is the case, it launches a function.
}
\usage{
checkIfAny(action = NULL, envirName = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{action}{
The function to be launched
}
  \item{envirName}{
The environment where the data sets are to be found
}
}

\author{
Marina Saez Andreu
}
