\name{checkIfAnyGapOrTs}
\alias{checkIfAnyGapOrTs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
checkIfAny: check if there are any time series or gap sets in the environment
}
\description{
This functions checks whether there are any time series or gap sets in the environment (KTSEnv). If it is the case, it launches a function.
}
\usage{
checkIfAnyGapOrTs(action = NULL, envirName = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{action}{
The function to be launched
}
  \item{envirName}{
The environment where the data sets are to be found
}
}


\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
