\name{checkIfAnyRm}
\alias{checkIfAnyRm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
checkIfAny: check if there is some recurrence matrix in the environment
}
\description{
This functions checks whether there are any recurrence matrices in the environment (KTSEnv). If it is the case, it launches a function.
}
\usage{
checkIfAnyRm(action = NULL, envirName = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{action}{
The function to be launched
}
  \item{envirName}{
The environment where the data sets are to be found
}
}


\author{
Marina Saez Andreu
}
