\name{checkIfAnyRmTs}
\alias{checkIfAnyRmTs}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
checkIfAnyRmTs: check if there is some recurrence matrix and some time series in the environment
}
\description{
This functions checks whether there are any recurrence matrices and time seriesin the environment (KTSEnv). If it is the case, it launches a function.
}
\usage{
checkIfAnyRmTs(action = NULL, envirName = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{action}{
The function to be launched
}
  \item{envirName}{
The environment where the data sets are to be found
}
}


\author{
Marina Saez Andreu
}
