\name{createCrossRMPlot}
\alias{createCrossRMPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createCrossRMPlot: creates a cross recurrence plot
}
\description{
This function creates a cross recurrence plot from a previously created cross recurrence matrix. It is used through the Plot Cross Recurrence Plot in the Plots Menu
}
%- maybe also 'usage' for other objects documented here.

\references{
Marwan,R., Romano, M.C., Thiel,M., Kurths,J.(2007): Recurrence plots for the analysis of complex systems. Physics Reports 438, 237-329.
}
\author{
Marina Saez Andreu
}
