\name{createDistMatrix}
\alias{createDistMatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createDistMatrix: creates distance matrix
}
\description{
This function calculates and plots a distance matrix via interface.
}

\details{
A distance matrix is an unthresholded recurrence matrix.The norm can be Euclidean or Infinity.  
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
