\name{createFANplot}
\alias{createFANplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createFANplot: creates a plot from a fixed amount of neighbors (FAN) recurrence matrix
}
\description{
This function creates a recurrence plot from a previously created fixed amount of neighbors (FAN) recurrence matrix.
}
%- maybe also 'usage' for other objects documented here.

\references{
Marwan,R., Romano, M.C., Thiel,M., Kurths,J.(2007): Recurrence plots for the analysis of complex systems. Physics Reports 438, 237-329.
}
\author{
Marina Saez Andreu
}
