\name{createGapRb}
\alias{createGapRb}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createGapRb: creates a radio button of gap sets
}
\description{
This function creates a radiobutton showing all the gap sets available
}
\usage{
createGapRb(labTitle = "Gap sets", envir = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labTitle}{
Title for the check box
}
  \item{envir}{
Environment
}
}

\value{
A variable to be passed to the corresponding OnOk function
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line

