\name{createJointRMPlot}
\alias{createJointRMPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createJointRMPlot: creates a cross recurrence plot
}
\description{
This function creates a joint recurrence plot from a previously created joint recurrence matrix. It is used through the Plot Joint Recurrence Plot in the Plots Menu
}
%- maybe also 'usage' for other objects documented here.

\references{
Marwan,R., Romano, M.C., Thiel,M., Kurths,J.(2007): Recurrence plots for the analysis of complex systems. Physics Reports 438, 237-329.
}
\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
