\name{createNote}
\alias{createNote}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createNote: creates a note on the input panel
}
\description{
This function creates a note on the input panel
}
\usage{
createNote(labTitle = NULL, pady = c(10, 10))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labTitle}{
Text
}
  \item{pady}{
Upper and lower margins
}
}

\author{
Marina Saez andreu
}
