\name{createOK}
\alias{createOK}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createOK: creates the OK or NEXT button on the input panel
}
\description{
This function creates the OK or NEXT button on the input panel.When the button is pressed the corresponding function is launched.
}
\usage{
createOK(labTitle = "NEXT", action = NULL, width = 7, panel = KTSEnv$subPanR4C1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labTitle}{
Usually, its values are "OK" or "NEXT"
}
  \item{action}{
Function that will be launched when the button is pressed.
}
  \item{width}{
Button width
}
  \item{panel}{
Panel were the button will be placed
}
}

\author{
Marina Saez Andreu
}
