\name{createSubPanR4C1}
\alias{createSubPanR4C1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createSubPanR4C1: creates the input panel on the main screen
}
\description{
This function creates a blank input panel on the main screen. It is used whenever is necessary to erase the elements of the input panel (title, check boxes, radiobutton etc.) and write new ones.It is used internally
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
