\name{createTitle}
\alias{createTitle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createTitle: creates a title in the input panel
}
\description{
This function creates (on the input panel) a label  that works as a second level title.It is used internally
}
\usage{
createTitle(labTitle = "Title")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labTitle}{
The text of the title
}
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
