\name{cumuKTS}
\alias{cumuKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
cumuKTS: creates a cumulated time series
}
\description{
This function creates a cumulated time series, that is, each point is the sum of all the previous values in the time series.It is particulary useful when the input time series has zero mean.It is used through the button Cumulative sum in the Time Series Menu
}

\value{
The cumulated time series
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
