\name{destroyMainScreen}
\alias{destroyMainScreen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to destroy KarsTS main screen
}
\description{
A function to destroy KarsTS main screen. It is normally destroyed via interface.
}
\usage{
destroyMainScreen()
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
