\name{destroyWelcome}
\alias{destroyWelcome}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
destroyWelcome: destroys Susana welcome screen
}
\description{
This function destroys Susana welcome screen. It is used through the button Start in the welcome screen.
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
