\name{determinismKTS}
\alias{determinismKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
determinismKTS: estimate determinism
}
\description{
This function is used to study the determinism of a system, based on the number and length of the diagonal lines that contains the recurrence matrix that represents the system. It is used through the button Determinism in the Anaylisis menu.
}

\value{
The following outputs are written on KarsTS output window: recurrence rate, determinism, ratio and summary of the lengths of the diagonal lines. Besides, a new window containing a histogram of diagonal lines pops up. 
}
\references{
Marwan,R., Romano, M.C., Thiel,M., Kurths,J.(2007): Recurrence plots for the analysis of complex systems. Physics Reports 438, 237-329.

}
\author{
Marina Saez Andreu
}
