\name{E1dAndE2d}
\alias{E1dAndE2d}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
E1dAndE2d: invariants E1(d) and E2(d)
}
\description{
This function uses the function nonlinearTseries::estimateEmbeddingDim to plot the invariants E1(d) and E2(d) (Cao,1997). E1(d) helps finding the embedding dimension of a scalar time series and E2(d) can distinguish stochastic from deterministic signals.It is used through the button E1d & E2d in the Plots Menu.
}

\references{
Cao, L (1997): Practical method for determining the minimum embedding dimension of a scalar time series. Physica D: Nonlinear Phenomena, 110,1, pp. 43-50.
}
\author{
Marina Saez Andreu
}
