\name{endingLines}
\alias{endingLines}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
endingLines: add three lines of asterisks on the output panel
}
\description{
This function adds three lines of asterisks on the output panel.It is used to separate outputs of different functions.It is used internally
}
\usage{
endingLines()
}

\author{
Marina Saez Andreu
}
