\name{exportall}
\alias{exportall}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
exportall: exports all types of data sets
}
\description{
This function is used to export time series, gap sets or recurrence matrices to csv or txt files. It is used through the button Export in the Time Series or Gap Sets menus. 
}
%- maybe also 'usage' for other objects documented here.
\details{
All types of data sets can be exported both from the Time Series  menu or from the Gap Sets menu.It is possible to export more than one time series to the same file; however, each gap set and recurrence matrix must be stored separatedly. The files are created in the woking directory. 
}

\author{
Marina Saez Andreu
}
