\name{fnnKTS}
\alias{fnnKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
fnn: plots embedding dimendion vs false nearest neighbors
}
\description{
This function plots embedding dimension vs false nearest neighbors using tseriesChaos::false.nearest and tseriesChaos::plot.false.nearest. It is used to find the minimum embedding dimension. It is used through the button FNN in the Plots Menu.
}

\author{
Marina Saez andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
