\name{gapCheckedTF}
\alias{gapCheckedTF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
gapCheckedTF: identifies checked gap sets
}
\description{
This function identifies which gap sets have been checked from a check box
}
\usage{
gapCheckedTF(prefix = "gcbValue", envir = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{prefix}{
A prefix to reconstruct the names that were assigned to the variables in the check box.
}
  \item{envir}{
The environment
}
}

\value{
A logical vector
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
