\name{gapDetect}
\alias{gapDetect}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
gapDetect: identifies the gap sets currently loaded
}
\description{
This function identifies the gap sets currently loaded in the environment KTSEnv
}
\usage{
gapDetect()
}
%- maybe also 'usage' for other objects documented here.

\value{
A vector containing the names of the gap sets
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
