\name{getCoordsKTS}
\alias{getCoordsKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getCoordsKTS: get coordinates from a plot
}
\description{
This function allows to manually select points from a plot and get their coordinates. It is used through the button Get coordinates in the Plots menu.
}
%- maybe also 'usage' for other objects documented here.
\details{
This function has two steps. In the first place, it is necessary to select a time series and plot it. The plot appears on a pop-up window. Then, the user clicks on one or more points and they turn green. When the user presses the Write results button, a list of coordinates (time and value) appears on the output window. 
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
