\name{getMaxNegSlope}
\alias{getMaxNegSlope}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getMaxNegSlope: get the maximum negative slope
}
\description{
This function returns the greater negative difference in a time series (greater in absolute value). It is used internally.
}
\usage{
getMaxNegSlope(timSerVals)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{timSerVals}{
The time series (only the values, not the times)
}
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
