\name{getMaxPosSlope}
\alias{getMaxPosSlope}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getMaxPosSlope: get the maximum negative slope
}
\description{
This function returns the greater positive difference in a time series.It is used internally.
}
\usage{
getMaxPosSlope(timSerVals)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{timSerVals}{
The time series (only the values, not the times)
}
}

\author{
Marina Saez Andreu
}
