\name{getOtherErrEstim}
\alias{getOtherErrEstim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getOtherErrEstim: get some error estimates from a linear fit
}
\description{
This function gets the relative root mean square error,the mean absolute error and bias error.
}
\usage{
getOtherErrEstim(observed, predicted)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{observed}{
Observed values
}
  \item{predicted}{
Predicted values
}
}


\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
