\name{getSamPerTable.1Freq}
\alias{getSamPerTable.1Freq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getSamPerTable.1Freq: get sampling periods table (time series with one frequency)
}
\description{
This function is a simplified version of getSamPerTable for time series that have a single sampling period. It is used internally.
}

\author{
Marina Saez Andreu
}
