\name{getScreenSize}
\alias{getScreenSize}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getScreenSize: gets the computer screen size
}
\description{
This function gets the computer screen size.It is used internally.
}
\usage{
getScreenSize()
}

\value{
The width and heigth of the screen are assigned to the enviroment susEnv.
}

\author{
Marina Saez Andreu
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
