\name{invariantsKTS}
\alias{invariantsKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
invariantsKTS: invariant plots
}
\description{
This function estimates the correlation sum and dimension and the KS Entropy through the functions nonlinearTseries::corrDim and nonlinearTseries::sampleEntropy. It returns three plots: correlation sum, correlation dimension and sample entropy.The first two appear on the same pop-up window.It is used internally through the button Invariants in the Analysis menu.
}

\author{
Marina Saez Andreu
}
