\name{KarsTS}
\alias{KarsTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
KarsTS: launches KarsTS
}
\description{
This function is used to launch KarsTS interface.
This is the only function in the package that is called through the R command window. The rest are used internally or manipulated via interface.If the user closes the interface accidentally, this function can be simply run again.  

}
\usage{
KarsTS(skipWelcome = FALSE)
}

\arguments{
  \item{skipWelcome}{
When TRUE, the interface main screen appears directly
}
}

\author{
Marina Saez Andreu
}
