\name{laminarityKTS}
\alias{laminarityKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
laminarityKTS: estimate laminarity
}
\description{
This function is used to study the laminarity of a system, based on the number and length of the vertical lines that contains the recurrence matrix that represents the system. It is used internally through the button Laminarity in the Anaylisis menu.
}

\value{
The following outputs are written on KarsTS output window: recurrence rate, laminarity, ratio and summary of the lengths of the vertical lines. Besides, a new window containing a histogram of vertical lines pops up. 
}
\references{
Marwan,R., Romano, M.C., Thiel,M., Kurths,J.(2007): Recurrence plots for the analysis of complex systems. Physics Reports 438, 237-329.

}
\author{
Marina Saez Andreu
}

