\name{linCorrKTS}
\alias{linCorrKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
linCorrKTS: linear correlation plot
}
\description{
This function plots the autocorrelation and partial autocorrelation functions or the cross correlation function, depending on the number of input time series. It is used internally through the Linear correlation button in the Plots menu.
}
%- maybe also 'usage' for other objects documented here.

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
