\name{linearityKTS}
\alias{linearityKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
linearityKTS: linearity tests
}
\description{
This function runs a number of linearity tests using the functions surrogateTest and nonlinearityTest from package nonlinearTseries.It is used internally through the Linearity button in the Analysis menu.
}

\author{
Marina Saez Andreu
}
