\name{littleTest}
\alias{littleTest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
littleTest: Little's MCAR test
}
\description{
This function runs the little MCAR test using the function BaylorEdPsych::LittleMCAR.It is used through the Little's MCAR button in the Analysis menu.
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
