\name{loadAllTypes}
\alias{loadAllTypes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
loadAllTypes: load all types of data sets
}
\description{
This function loads all types of data sets (time series, data sets and recurrence matrices). It generates two buttons: one is used to load R files and the other is used to import csv and txt files. The function automatically recognizes the type of data set, checks whether the format is right and places the data sets in the environment susEnv. It is used through the Load button in the Time Series or the Gap Sets menus.
}

\author{
Marina Saez Andreu
}
