\name{mainScreen}
\alias{mainScreen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mainScreen: launches the main screen
}
\description{
This function launches KarsTS main screen when the welcome screen is destroyed.It is used internally.
}
\usage{
mainScreen()
}

\author{
Marina Saez Andreu
}
