\name{mergeTsOrGap}
\alias{mergeTsOrGap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mergeTsOrGap: merges time series or gap sets
}
\description{
This function merges two or more time series and two or more gap sets. The time series cannot overlap and must have the same unique sampling period. The gap sets must coome from time series with the same initial date and sampling period.It is used through the button Merge of the Time Series or the Gap Sets menus.
}

\value{
The merged time series, the merged gap set or both.
}

\author{
Marina Saez Andreu
}
