\name{modeKTS}
\alias{modeKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
modeKTS: computes the mode
}
\description{
This function computes the mode after removing missing values.
}
\usage{
modeKTS(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The data 
}
}


\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
