\name{mutInf}
\alias{mutInf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mutInf: mutual information
}
\description{
This function calculates the mutual information between two time series via interface.The core functions belong to the package \code{infotheo}. 
}

\author{
Marina Saez Andreu
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
