\name{myLinModel}
\alias{myLinModel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
myLinModel: performs a linear fit
}
\description{
This function is used to perform a linear model between observed and predicted values. It also prepares the outputs to be writen on the output window and plots different  graphics.It is the core of the goodnessFilling function.
}
\usage{
myLinModel(observed, predicted)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{observed}{
Observed values
}
  \item{predicted}{
Predicted values
}
}

\author{
Marina Saez Andreu
}
