\name{NAs4Resamp}
\alias{NAs4Resamp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
NAs4Resamp: upsamples a time series.
}
\description{
NAs4Resamp introduces NAs in a time series for upsampling, via interface.The user needs to provide the new sampling period in seconds.This function only introduces NAs, then they need to be filled (menu Filling methods).
}
\value{
The upsampled time series in KarsTS format.
}
\author{
Marina Saez Andreu
}
