\name{packagesToImport}
\alias{packagesToImport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
packagesToImport: packages to import to Susana
}
\description{
This function imports from other packages the functions that Susana needs.It is used internally.
}
\usage{
packagesToImport()

}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
