\name{pcaKTS}
\alias{pcaKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
pcaKTS: principal component analysis
}
\description{
This function performs a principal component analysis using the function prcomp.It is used through the button PCA in the Analysis menu.
}

\value{
The punctuations as new time series in the environment KTSEnv. A summary on the output window, including the loadings and the variance explained.
}

\author{
Marina Saez Andreu
}
