\name{removeAllTypes}
\alias{removeAllTypes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
removeAllTypes: removes data sets
}
\description{
This function removes all types of data sets (time series, gap sets, recurrence matrices) from the environment susEnv. It is used through the button Remove in the Time Series or Gap Sets menus.
}

\author{
Marina Saez Andreu
}
