\name{renameAllTypes}
\alias{renameAllTypes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
renameAllTypes: rename all types of data sets
}
\description{
This function is used to rename time series, gap sets or recurrence matrices. The data set with the original name is kept. It is used through the Rename button in the Time Series or the Gap Sets menus.
}

\author{
Marina Saez Andreu
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
