\name{rmCheckedTF}
\alias{rmCheckedTF}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rmCheckedTF: identifies checked recurrence matrices
}
\description{
This function identifies which time series have been checked from a check box.
}
\usage{
rmCheckedTF(prefix = "rcbValue", envir = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{prefix}{
A prefix to reconstruct the names that were assigned to the variables in the check box.
}
  \item{envir}{
The environment
}
}

\value{
A logical vector
}

\author{
Marina Saez Andreu
}
