\name{rollStatisticsKTS}
\alias{rollStatisticsKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rollStatisticsKTS: rolling statistics
}
\description{
This function gathers, via interface, the outputs that are necessary to run the function getRollStatistics. It is used through the button Rolling statistics in the Analysis menu.
}

\author{
Marina Saez Andreu
}
