\name{roundKTS}
\alias{roundKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
roundKTS: rounding
}
\description{
This function rounds the values of one or more time series to a number of decimal places or significant digits. When the user enters both, the significant digits option takes priority.It is used through the button Round in the Time Series menu.
}

\value{
The rounded time series in the environment KTSEnv.
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
