\name{RPKTS}
\alias{RPKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Recurrence of probability
}
\description{
RPKTS:This function is used to calculate the recurrece probability via interface
}

%- maybe also 'usage' for other objects documented here.
\details{
When two time series are selected, KarsTS calculates their correlation probability of recurrence too.
}

\references{
Romano, M. C. (2004). Synchronization Analysis by Means of Recurrences in Phase Space, Universitat Postdam. Doctoral dissertation.
}
\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
