\name{saveAllTypes}
\alias{saveAllTypes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
saveAllTypes: saves all types of data sets.
}
\description{
This function saves all types of data sets (time series, gap sets or recurrence matrices) to one or more R files in the working directory. When various data sets are to be saved to the same file, the user must provide name; otherwise, the file bear the name of the data set.Through the button save in the Time Series or the Gap Sets menu. 
}

\author{
Marina Saez Andreu
}
