\name{scaleKTS}
\alias{scaleKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
scaleKTS: scale time series
}
\description{
This function is used to scale or more time series. It handles the inputs and outputs of the myScale function.Through the button Scale in the Time Series menu.
}
\author{
Marina Saez Andreu
}
