\name{setwdKTS}
\alias{setwdKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
setwdKTS: set working directory
}
\description{
This function is used to change the working directory by means of the Set WD button on the main screen.
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
