\name{showHelp}
\alias{showHelp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
showHelp: shows Susana help
}
\description{
This function is used to open a short help file through the Help button on Susana main screen. For further help, a User's Guide in pdf format is available.
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
