\name{stationarityKTS}
\alias{stationarityKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
stationarityKTS: stationarity tests
}
\description{
This function runs a number of stationarity tests using the functions PP.test and Box.test from package stats and adf.test and kpss.test from package tseries.Through the Stationarity button in the Analysis menu.
}

\author{
Marina Saez Andreu
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
