\name{statisticsKTS}
\alias{statisticsKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
statisticsKTS: calculates the statistics of one or more time series
}
\description{
This function handles the inputs and outputs from the function getStatistics, which is used to calculate the statistics of a time series.Through the button Statistics in the Analysis menu.
}

\value{
A table on the output window.
}

\author{
Marina Saez Andreu
}
