\name{theilerKTS}
\alias{theilerKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
theilerKTS: apply Theiler's window via interface
}
\description{
It applies a Theiler's window to a recurrence matrix via interface.
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
