\name{verifyDateEntry}
\alias{verifyDateEntry}
\title{
verifyDateEntry: verify date entries
}
\description{
This function checks whether a date is valid or not.It has six inputs (seconds, minutes,hours,day,month and year); these come from a set of text entries.It is used internally.
}
\usage{
verifyDateEntry(valSecs, valMins, valHour, valDay, valMonth, valYear)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{valSecs}{
Seconds
}
  \item{valMins}{
Minutes
}
  \item{valHour}{
Hour
}
  \item{valDay}{
Day
}
  \item{valMonth}{
Month
}
  \item{valYear}{
Year
}
}

\author{
Marina Saez Andreu
}

