\name{welcomeScreen}
\alias{welcomeScreen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
welcomeScreen: launches the welcome screen
}
\description{
This function launches Susana welcome screen. It is used internally.
}
\usage{
welcomeScreen()
}
%- maybe also 'usage' for other objects documented here.

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
