\name{windRoseKTS}
\alias{windRoseKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
windRoseKTS: wind rose
}
\description{
This function plots a wind rose from directional data. The data must be in sexagesimal degrees. It is used through the button Wind Rose in the Analysis menu.
}

\author{
Marina Saez Andreu
}
