\name{writeMethodTitle}
\alias{writeMethodTitle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
writeMethodTitle: writes a title on Susana output window
}
\description{
This function writes a title on Susana output window. It is used internally.
}
\usage{
writeMethodTitle(titleMethod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{titleMethod}{
The text of the title
}
}

\author{
Marina Saez Andreu
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
