\name{buttons1}
\alias{buttons1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
buttons1: create the buttons corresponding to the Time Series Menu
}

\description{
This function creates the buttons corresponding to the Time Series Menu
}

\author{
Marina Saez Andreu
}
