\name{createGapChb}
\alias{createGapChb}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createGapChb: creates a check box of gap sets
}
\description{
This function creates a check box showing all the gap sets available. It is used internally
}
\usage{
createGapChb(labTitle = "Gap sets", envir = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labTitle}{
Title for the check box
}
  \item{envir}{
Environment
}
}

\value{
Variables to be passed to the corresponding OnOk function
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
