\name{createRb}
\alias{createRb}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createRb: creates a radio button
}
\description{
This function creates a radio button
}
\usage{
createRb(variable = NULL, dataVector = NULL, panel = KTSEnv$subPanR4C1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{variable}{
The variable to which the radiobutton is linked
}
  \item{dataVector}{
The names of the buttons
}
  \item{panel}{
The panel where to place the radiobutton (defaults to the main screen input panel)
}
}

\author{
Marina Saez Andreu
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
