\name{createSpecGaps}
\alias{createSpecGaps}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createSpecGaps: creates a specific gap in a time series
}
\description{
This function creates a gap between two dates chosen by the user.It is used through the Specific Gaps button in the Gap Sets menu.
}
\value{
The function creates a gap set and applies it to a copy of the input time series.Both the gap set and the new time series appear in the environment susEnv. 
}

\author{
Marina Saez Andreu
}
