\name{createTITLE}
\alias{createTITLE}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
createTITLE: create a title on the input panel
}
\description{
This function creates (on the input panel) a label  that works as a first level title.
}
\usage{
createTITLE(labTitle = "TITLE", panel = KTSEnv$subPanR4C1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{labTitle}{
Text of the title
}
  \item{panel}{
Panel where the title is to be written (defaults to the input panel on the main screen)
}
}

\author{
Marina Saez Andreu
}
