\name{destroyMainScreen}
\alias{destroyMainScreen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to destroy KarsTS main screen
}
\description{
A function to destroy KarsTS main screen. It is normally destroyed via interface.
}
\usage{
destroyMainScreen()
}

\author{
Marina Saez Andreu
}
