\name{destroyWelcome}
\alias{destroyWelcome}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
destroyWelcome: destroys KarsTS welcome screen
}
\description{
This function destroys KarsTS welcome screen. It is used through the button Start in the welcome screen.
}

\author{
Marina Saez Andreu
}
