\name{gapDetect}
\alias{gapDetect}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
gapDetect: identifies the gap sets currently loaded
}
\description{
This function identifies the gap sets currently loaded in the environment KTSEnv
}
\usage{
gapDetect()
}
%- maybe also 'usage' for other objects documented here.

\value{
A vector containing the names of the gap sets
}

\author{
Marina Saez Andreu
}
