\name{getClassEnvir}
\alias{getClassEnvir}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getClassEnvir: get objects of a class from an environment
}
\description{
This function identifies objects of a class in an environment 
}
\usage{
getClassEnvir(classGet = "list", envir = KTSEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{classGet}{
The class
}
  \item{envir}{
The environment
}
}

\value{
A vector with the names of the objects
}

\author{
Marina Saez Andreu
}