\name{getStatistics}
\alias{getStatistics}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getStatistics: calculates the statistics of a time series
}
\description{
This function calculates the minimum, first quartile, median, mean, third quartile, maximum and standard deviation of a time series.It is used internally.
}
\usage{
getStatistics(selTs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{selTs}{
The time series
}
}

\value{
A vector containing the statistics values
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
