\name{getUniqueSampPer}
\alias{getUniqueSampPer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
getUniqueSampPer: get unique sampling periods
}
\description{
This function get the time steps existing in a time series and sorts them according to the number of times they appear (starting by the most repeated). It returns a table with this information.It is used internally.
}
\usage{
getUniqueSampPer(timeSer)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{timeSer}{
The time series
}
}

\author{
Marina Saez Andreu
}
