\name{goodnessFilling}
\alias{goodnessFilling}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
goodnessFilling: estimate the goodness of the filling
}
\description{
This function performs a linear fit between observed and predicted values. The predicted values come from an artificial gap set that has been filled. When true missing values are imputed, it is impossible to estimate the goodness of the filling. It is used through the Check filling button in the Filling menu.
}
\value{
A summary of the fit in the output window and different plots in pop-up windows.
}

\author{
Marina Saez Andreu
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
