\name{histKTS}
\alias{histKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
histKTS: plots histogram
}
\description{
This function plots the histogram of a time series values. Its inputs are the time series and, optionally, the approximate number of bars. It is used internally through the button Histogram in the Plots menu
}
\author{
Marina Saez Andreu
}
