\name{loadKarsTSFonts}
\alias{loadKarsTSFonts}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
loadKarsTSFonts: load KarsTS fonts
}
\description{
This function creates fonts for KarsTS and assign them to the environment susEnv.It is used internally.
}
\usage{
loadKarsTSFonts()
}
%- maybe also 'usage' for other objects documented here.

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
