\name{missForestKTS}
\alias{missForestKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
missForestKTS: filling missing values with random forest algorithm
}
\description{
This function is used to fill missing values using the function missForest::missForest.The user must choose the input and output time series, the number of trees and the maximum number of iterations. It is used through the button MissForest in the Filling menu.
}

\value{
The filled time series in the environment KTSEnv and a summary of the results on the output window. 
}

\author{
Marina Saez Andreu
}
