\name{modeKTS}
\alias{modeKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
modeKTS: computes the mode
}
\description{
This function computes the mode after removing missing values.
}
\usage{
modeKTS(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The data 
}
}


\author{
Marina Saez Andreu
}
