\name{mutInf}
\alias{mutInf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
mutInf: mutual information
}
\description{
This function calculates the mutual information between two time series via interface.The core functions belong to the package \code{infotheo}. 
}

\author{
Marina Saez Andreu
}

