\name{naApproxKTS}
\alias{naApproxKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
naApproxKTS: filling by linear interpolation
}
\description{
This function fills missing values in a time series by means of linear interpolation.It is used through the Linear button in the Filling menu.
}

\value{
The filled time series in the environment KTSEnv and a summary of the results on the output window.
}

\author{
Marina Saez Andreu
}
