\name{normalityKTS}
\alias{normalityKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
normalityKTS: normality tests
}
\description{
This function runs a number of normality tests using the functions uniNorm, mardiaTest and hzTest from package MVN. It is used through the Normality button in the Analysis menu.
}

\author{
Marina Saez Andreu
}

