\name{packagesToImport}
\alias{packagesToImport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
packagesToImport: packages to import to KarsTS
}
\description{
This function imports from other packages the functions that KarsTS needs.It is used internally.
}
\usage{
packagesToImport()

}

\author{
Marina Saez Andreu
}
