\name{plotTimeSeries}
\alias{plotTimeSeries}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
plotTimeSeries: plot time series
}
\description{
This function is used to plot one or more time series. They can be plotted using lines, points or both. The plot can appear on the plot window (main screen) or in a new window. In the second case, it is possible to select a part of the time series and plot it in another window.It is used through the Plot ts button in the Plots menu.
}

\author{
Marina Saez Andreu
}
