\name{rmDetect}
\alias{rmDetect}
%- Also NEED an '\alias' for EACH other topic documented here.
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
rmDetect: identifies the recurrence matrices currently loaded
}
\description{
This function identifies the recurrence matrices currently loaded in the environment susEnv. It is used internally.
}
\usage{
rmDetect()
}
%- maybe also 'usage' for other objects documented here.

\value{
A vector containing the names of the recurrence matrices
}

\author{
Marina Saez Andreu
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
