\name{saveReport}
\alias{saveReport}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
saveReport: saves report
}
\description{
This function saves the contents of the output window to a txt file in the working directory.The file name is report plus the date of creation.Through the button Save report to txt file (lower part of the main screen).
}

\author{
Marina Saez Andreu
}
