\name{scaleKTS}
\alias{scaleKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
scaleKTS: scale time series
}
\description{
This function is used to scale or more time series. It handles the inputs and outputs of the myScale function.Through the button Scale in the Time Series menu.
}
\author{
Marina Saez Andreu
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
