\name{scattTimeSeries}
\alias{scattTimeSeries}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function for plotting time series in the phase space
}
\description{
This function plots time series in a two dimensional or three dimensional phase space. The time series can be embedded so that the sum of the dimension of all time series is 2 or 3. Is is used through the Phase Portraits button in the Plots Menu. 
}


\author{
Marina Saez Andreu
}
