\name{selectionTS}
\alias{selectionTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
selectionTS
}
\description{
This function is used to cut a piece of time series, resample a time series or both. The user chooses the initial and final dates of the new time series and the resampling factor.The dates default to the original time series extremes and the resampling factor defaults to 1(no resampling). Through the Cut and resampling button in the Time Series menu.
}

\author{
Marina Saez Andreu
}
