\name{setwdKTS}
\alias{setwdKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
setwdKTS: set working directory
}
\description{
This function is used to change the working directory by means of the Set WD button on the main screen.
}

\author{
Marina Saez Andreu
}
