\name{showHelp}
\alias{showHelp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
showHelp: shows KarsTS help
}
\description{
This function is used to open a short help file through the Help button on KarsTS main screen. For further help, a User's Guide in pdf format is available.
}

\author{
Marina Saez Andreu
}
