\name{stinemannKTS}
\alias{stinemannKTS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
stinemannKTS: filling by Stinemann's interpolation
}
\description{
This function fills missing values in a time series by means of Stinemann's interpolation.Through the Stinemann's button in the Filling menu.
}
\value{
The filled time series in the environment KTSEnv and a summary of the results on the output window.
}

\author{
Marina Saez Andreu
}

