\name{tsDetect}
\alias{tsDetect}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
tsDetect: identifies the time series currently loaded
}
\description{
This function identifies the time series currently loaded in the environment susEnv
It is used internally
}
\usage{
tsDetect()
}
%- maybe also 'usage' for other objects documented here.

\value{
A vector containing the names of the time series
}

\author{
Marina Saez Andreu
}
