\name{welcomeScreen}
\alias{welcomeScreen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
welcomeScreen: launches the welcome screen
}
\description{
This function launches KarsTS welcome screen. It is used internally.
}
\usage{
welcomeScreen()
}
%- maybe also 'usage' for other objects documented here.

\author{
Marina Saez Andreu
}
