\name{writeMethodTitle}
\alias{writeMethodTitle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
writeMethodTitle: writes a title on KarsTS output window
}
\description{
This function writes a title on KarsTS output window. It is used internally.
}
\usage{
writeMethodTitle(titleMethod)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{titleMethod}{
The text of the title
}
}

\author{
Marina Saez Andreu
}
